purchase_product

+++ app/Http/Controllers/Slider/SliderController.php
@@ -3,11 +3,9 @@
 namespace App\Http\Controllers\Slider;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\SliderRequest;
-use Illuminate\Http\Request;
-use App\Models\PhotoSliders;
+use App\Http\Requests\Slider\SliderRequest;
+use App\Models\Sliders;
 use Illuminate\Support\Facades\Storage;
-use Illuminate\Support\Facades\Validator;
 
 class SliderController extends Controller
 {
@@ -16,8 +14,9 @@ class SliderController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index() {
-        $sliders = PhotoSliders::paginate(10);
+    public function index()
+    {
+        $sliders = Sliders::paginate(10);
         return view('pages.manageSlider.slider_index', ['sliders' => $sliders]);
     }
 
@@ -26,53 +25,39 @@ class SliderController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create() {
+    public function create()
+    {
         return view('pages.manageSlider.slider_create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \App\Http\Requests\SliderRequest  $request
+     * @param  \App\Http\Requests\Slider\SliderRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(SliderRequest $request) {
+    public function store(SliderRequest $request)
+    {
         $data = $request->validated();
-        $title = $request->title;
-        $description = $request->description;
-        $slider = $request->file('slider')->getClientOriginalName();
-        $type = $request->file('slider')->getMimeType();
-        $size = $request->file('slider')->getSize();
-        $validate = Validator::make($data,[$title,$description,$slider,$type,$size]);
-        if($validate->fails()) {
-            return redirect()::back()->withInput()->withErrors($validate->errors());
-        } else {
-            $sliders = new PhotoSliders();
-            $sliders->title = $title;
-            $sliders->description = $description;
-            $sliders->name = $slider;
-            $sliders->size = $size;
-            $sliders->type = $type;
-            $sliders->save();
-            $request->file('slider')->storeAs('public/images/slider',$slider);
-            return back();
-        }
-       /* try
-        {
-            $sliders->save();
-            $request->file('slider')->storeAs('public/images/slider',$data['slider']->getClientOriginalName());
-            $success = 'Post slider success';
-
+        $sliders = new Sliders();
+        $slider = $data['slider']->getClientOriginalName();
+        $sliderName = time().'_'.$slider;
+        try {
+            $sliders->create([
+                'title' => $data['title'],
+                'description' => $data['description'],
+                'name' => $sliderName
+            ]);
+            $data['slider']->storeAs('public/images/slider', $sliderName);
+            $success = 'Add slider success';
             return redirect()
                 ->route('slider.index')
-                ->with('success',$success);
-        }
-        catch (\Exception $e)
-        {
+                ->with('success', $success);
+        } catch (\Exception $e) {
             \Log::error($e);
-            $error = 'Post slider failed';
+            $error = 'Add slider fail';
         }
-        return back()->with('error',$error);*/
+        return back()->with('error',$error);
     }
 
     /**
@@ -81,7 +66,8 @@ class SliderController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id) {
+    public function show($id)
+    {
         //
     }
 
@@ -91,8 +77,9 @@ class SliderController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id) {
-        $slider = PhotoSliders::findOrFail($id);
+    public function edit($id)
+    {
+        $slider = Sliders::findOrFail($id);
         return view('pages.manageSlider.slider_edit', [
             'slider' => $slider,
         ]);
@@ -101,63 +88,33 @@ class SliderController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Http\Requests\Slider\SliderRequest  $request
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id) {
-        $slider = PhotoSliders::findOrFail($id);
-        $titleOld = $slider->title;
-        $descriptionOld = $slider->description;
-        $titleNew = $request->title;
-        $descriptionNew = $request->description;
-
-        /*Case 1:Update photo slider - Not update title,description*/
-        if ($titleNew == null && $descriptionNew == null) {
-            $sizeNew = $request->file('slider')->getSize();
-            $typeNew = $request->file('slider')->getMimeType();
-            $imageNew = $request->file('slider')->getClientOriginalName();
-            $request->file('slider')->storeAs('public/images/slider', $imageNew);
-            $titleNew = $titleOld;
-            $descriptionNew = $descriptionOld;
-            try
-            {
-                $msgSuccess = 'Update slider success';
-                $slider->title = $titleNew;
-                $slider->description = $descriptionNew;
-                $slider->name = $imageNew;
-                $slider->size = $sizeNew;
-                $slider->type = $typeNew;
-                $slider->save();
-                return redirect()
-                    ->route('slider.index')
-                    ->with('success', $msgSuccess);
-            } catch (\Exception $e) {
-                \Log::error($e);
-            }
-            $msgFail = 'Update slider fail';
+    public function update(SliderRequest $request, $id)
+    {
+        $data = $request->validated();
+        $sliders = Sliders::findOrFail($id);
+        $sliderImgOld = Storage::files('public/images/slider',$sliders->name);
+        $sliderImgNew = time().'_'.$data['slider']->getClientOriginalName();
+        $values = [
+            'title' => $data['title'],
+            'description' => $data['description'],
+            'name' => $sliderImgNew
+        ];
+        try {
+            $sliders->update($values);
+            $success = 'Update slider success';
+            $data['slider']->storeAs('public/images/slider', $sliderImgNew);
+            Storage::delete($sliderImgOld);
             return redirect()
                 ->route('slider.index')
-                ->with('error', $msgFail);
-        };
-        /*Case 2:Update title,description - Dont update photo*/
-        if ($request->file('slider') == null) {
-            try
-            {
-                $msgSuccess = 'Update slider success';
-                $slider->title = $titleNew;
-                $slider->description = $descriptionNew;
-                $slider->save();
-                return redirect()
-                    ->route('slider.index')
-                    ->with('success', $msgSuccess);
-            } catch (\Exception $e) {
-                \Log::error($e);
-            }
-            $msgFail = 'Update slider fail';
-            return redirect()
-                ->route('slider.index')
-                ->with('error', $msgFail);
+                ->with('success',$success);
+        } catch (\Exception $e) {
+            \Log::error($e);
+            $error = 'Update slider failed';
+            return back()->with('error',$error);
         }
     }
 
@@ -167,20 +124,22 @@ class SliderController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id) {
-        $slider = PhotoSliders::findOrFail($id);
+    public function destroy($id)
+    {
+        $slider = Sliders::findOrFail($id);
         try
         {
             $slider->delete();
-            Storage::delete('public/images/slider/' . $slider->name);
+            if(Storage::exists('public/images/slider/'.$slider->name)) {
+                Storage::delete('public/images/slider/'.$slider->name);
+            }
             $success = 'Delete slider success';
             return redirect()->route('slider.index')
                 ->with('success', $success);
         } catch (\Exception $e) {
             \Log::error($e);
-            $error = 'Delete slider fail';
+            $error = 'Delete slider failed';
         }
-        return redirect()->route('slider.index')
-            ->with('error', $error);
+        return back()->with('error', $error);
     }
 }
